app.py
import time
import pinecone
import pandas as pd
from transformers import AutoModelForCausalLM, AutoTokenizer, AutoModelForSeq2SeqLM
from langchain.llms import HuggingFaceLLM
from langchain.memories import ConversationBufferWindowMemory
from langchain.agents import Tool, initialize_agent
from langchain.chains import RetrievalQA
from langchain.retrievers import BaseRetriever
from tqdm.auto import tqdm

# Your Pinecone API key and other sensitive information should be stored securely,
# e.g., in environment variables or a configuration file.
PINECONE_API_KEY = 'your_pinecone_api_key_here'
pinecone.init(api_key=PINECONE_API_KEY, environment='us-west1-gcp')

# Function definitions (e.g., for initializing Pinecone, embedding documents, etc.)

def main():
    # Initialize Pinecone
    index_name = "langchain-retrieval-agent"
    # Ensure index creation and embedding logic here

    # Initialize the chatbot components (LLM, Memory, Retriever, etc.)

    # Running the chat loop
    while True:
        user_input = input("You: ")
        if user_input.lower() == "quit":
            break
        response = agent.process_input(user_input)
        print(f"AI: {response['text']}")

if __name__ == "__main__":
    main()

python app.py