requirements.txt
pinecone-client==0.10.1
transformers==4.18.0
pandas==1.3.5
tqdm==4.64.0